# Added heuristic using weighted keyword dictionary